+++ settfex/__init__.py
@@ -19,7 +19,7 @@ Usage:
     >>> asyncio.run(main())
 """
 
-__version__ = "0.5.0"
+__version__ = "0.6.0"
 __author__ = "batt"
 __license__ = "MIT"