+++ settfex/utils/__init__.py
@@ -4,6 +4,7 @@ from settfex.utils.data_fetcher import AsyncDataFetcher, FetcherConfig, FetchRes
 from settfex.utils.logging import get_logger, setup_logger
 from settfex.utils.session_cache import SessionCache, get_global_cache
 from settfex.utils.session_manager import SessionManager, get_shared_session
+from settfex.utils.youtube_transcript import fetch_youtube_transcript
 
 __all__ = [
     "AsyncDataFetcher",
@@ -15,4 +16,5 @@ __all__ = [
     "get_global_cache",
     "SessionManager",
     "get_shared_session",
+    "fetch_youtube_transcript",
 ]